renamed handlers package to routes
