+++ src/main/java/com/github/onsdigital/thetrain/App.java
@@ -8,13 +8,13 @@ import com.github.onsdigital.thetrain.exception.handler.PublishExceptionHandler;
 import com.github.onsdigital.thetrain.filters.AfterFilter;
 import com.github.onsdigital.thetrain.filters.BeforeFilter;
 import com.github.onsdigital.thetrain.filters.QuietFilter;
-import com.github.onsdigital.thetrain.handlers.AddFileToTransaction;
-import com.github.onsdigital.thetrain.handlers.CommitTransaction;
-import com.github.onsdigital.thetrain.handlers.GetTransaction;
-import com.github.onsdigital.thetrain.handlers.OpenTransaction;
-import com.github.onsdigital.thetrain.handlers.RollbackTransaction;
-import com.github.onsdigital.thetrain.handlers.SendManifest;
-import com.github.onsdigital.thetrain.handlers.VerifyTransaction;
+import com.github.onsdigital.thetrain.routes.AddFileToTransaction;
+import com.github.onsdigital.thetrain.routes.CommitTransaction;
+import com.github.onsdigital.thetrain.routes.GetTransaction;
+import com.github.onsdigital.thetrain.routes.OpenTransaction;
+import com.github.onsdigital.thetrain.routes.RollbackTransaction;
+import com.github.onsdigital.thetrain.routes.SendManifest;
+import com.github.onsdigital.thetrain.routes.VerifyTransaction;
 import com.github.onsdigital.thetrain.helpers.FileUploadHelper;
 import com.github.onsdigital.thetrain.response.JsonTransformer;
 import com.github.onsdigital.thetrain.service.PublisherService;
