+++ src/main/java/com/github/onsdigital/thetrain/routes/BaseHandler.java
@@ -1,4 +1,4 @@
-package com.github.onsdigital.thetrain.handlers;
+package com.github.onsdigital.thetrain.routes;
 
 import com.github.onsdigital.thetrain.exception.BadRequestException;
 import com.github.onsdigital.thetrain.json.request.Manifest;
